requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
sqlalchemy>=1.4.0
pyyaml>=6.0.0
stripe>=5.0.0
python-multipart>=0.0.5
pydantic[email]>=1.10.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0
